Cleanup

+++ src/jvmMain/kotlin/com/sdercolin/vlabeler/io/Wave.kt
@@ -59,7 +59,7 @@ suspend fun loadSampleChunk(
         Log.debug("Loading chunk $chunkIndex: offset=$offset")
         val readSize = stream.readNBytes(buffer, 0, chunkSize * frameSize)
         yield()
-        val readFrameCount = (readSize / frameSize).toInt()
+        val readFrameCount = readSize / frameSize
         for (frameIndex in 0 until readFrameCount) {
             val frameStart = frameIndex * frameSize
             val frameEnd = frameStart + frameSize
